requirements.txt
anthropic>=0.39.0
openai>=1.54.0
google-generativeai>=0.8.3
python-dotenv>=1.0.0
